Move GenericInput into helper-plugin
